fix(runtime): close dogfood completion gates
Fail closed when enforced USD budgets rely on backends that do not report cost, and preserve mid-turn budget stops as blocked outcomes.

Move the accounting tests into their same-file unit-test home, sanitize the Claude backend credential from Codex-only server launches, and record the fresh production replay evidence.
